pipeline/requirements.txt
# Core audio + ASR stack (fully local, CPU-friendly)
faster-whisper>=1.0.3
librosa>=0.10.2
soundfile>=0.12.1
pydub>=0.25.1
noisereduce>=3.0.2

# VAD (no downloads required)
webrtcvad-wheels>=2.0.14

# Diarisation fallback + metadata + trained romanised-Indic classifier
scikit-learn>=1.4.2
numpy>=1.26.0
scipy>=1.11.0

# Language detection (optional local model — download via download_models.py)
fasttext-wheel>=0.9.2

# CLI + utilities
tqdm>=4.66.0
jsonschema>=4.0.0
deepgram-sdk>=3.7.0
google-cloud-speech>=2.31.0

# Model hub (used by download_models.py only, not at inference time)
huggingface_hub>=0.22.0

# Optional: parquet / jsonl batch output
pandas>=2.2.0
pyarrow>=15.0.0

# Optional: pyannote.audio diarisation backend (needs HuggingFace token + accepted terms)
# Uncomment to enable:
# pyannote.audio>=3.1.1
# torch>=2.0.0
